refactor: decompose 3 oversized tools crate modules (#436)
Split the three largest production modules in crates/tools/src/ into
subdirectory modules with clear separation of concerns:

- git_tool.rs (558 LOC) -> git_tool/{mod,types,operations}.rs
- ask_user_question.rs (412 LOC) -> ask_user_question/{mod,types,tui_prompt,cli_prompt}.rs
- web_fetch.rs (528 LOC) -> web_fetch/{mod,types,cache,http}.rs

Public API unchanged -- all re-exports preserved via mod.rs.
All 37 module tests pass, clippy clean.
