+++ crates/tools/src/web_fetch_tests.rs
@@ -1,4 +1,5 @@
 use super::*;
+use crate::web_fetch_domains::PRE_APPROVED_DOMAINS;
 use crate::Tool;
 use futures::StreamExt;
